+++ build.gradle
@@ -28,7 +28,7 @@ dependencies {
     compile group: 'com.vimalselvam', name: 'cucumber-extentsreport', version: '3.0.2'
     compile group: 'net.masterthought', name: 'cucumber-reporting', version: '4.6.0'
     compile group: 'log4j', name: 'log4j', version: '1.2.17'
-    testCompile 'io.rest-assured:rest-assured:3.0.2'
+    compile 'io.rest-assured:rest-assured:3.0.2'
 }
 
 // Task to define arguments and configurations to properties file
