+++ src/main/java/salesforce/ui/pages/abstracts/HomePageAbstract.java
@@ -1,7 +1,7 @@
 package salesforce.ui.pages.abstracts;
 
 import salesforce.ui.BasePage;
-import salesforce.ui.pages.abstracts.account.AccountPageAbstract;
+import salesforce.ui.pages.account.abstracts.AccountPageAbstract;
 import salesforce.ui.pages.abstracts.task.NewTaskAbstract;
 import salesforce.ui.pages.campaign.abstracts.CampaignPageAbstract;
