+++ src/main/java/salesforce/ui/pages/account/abstracts/NewAccountPageAbstract.java
@@ -11,7 +11,7 @@
  *
  */
 
-package salesforce.ui.pages.abstracts.account;
+package salesforce.ui.pages.account.abstracts;
 
 import core.utils.StrategySetter;
 import salesforce.entities.Account;
