+++ src/main/java/salesforce/ui/pages/account/abstracts/OneAccountAbstract.java
@@ -1,5 +1,5 @@
 /*
- * @(#) AccountPageAbstract.java Copyright (c) 2019 Jala Foundation.
+ * @(#) OneAccountAbstract.java Copyright (c) 2019 Jala Foundation.
  * 2643 Av Melchor Perez de Olguin, Colquiri Sud, Cochabamba, Bolivia.
  * All rights reserved.
  *
@@ -11,27 +11,29 @@
  *
  */
 
-package salesforce.ui.pages.abstracts.account;
+package salesforce.ui.pages.account.abstracts;
 
 import salesforce.ui.BasePage;
 
 /**
- * AccountPageAbstract.
+ * OneAccountAbstract.
  * @author Luis Guardia.
  * @version 0.0.1
  */
-public abstract class AccountPageAbstract extends BasePage {
+public abstract class OneAccountAbstract extends BasePage {
 
     /**
-     * Click in New Account button.
-     * @return clickNewAccountBtn.
+     * Implement verifyComponentsAccount.
      */
-    public abstract NewAccountPageAbstract clickNewAccountBtn();
+    public abstract boolean verifyComponentsAccount();
 
     /**
-     * Implement: check the account list.
-     * @param name string.
-     * @return boolean.
+     * Implement get name to assert with steps.
      */
-    public abstract boolean checkAccountList(String name);
+    public abstract String getNameAccount();
+
+    /**
+     * Delete Account of Salesforce with specific name.
+     */
+    public abstract void deleteAccount();
 }
