+++ src/main/java/salesforce/ui/pages/account/classic/AccountClassicPage.java
@@ -11,13 +11,13 @@
  *
  */
 
-package salesforce.ui.pages.classic.account;
+package salesforce.ui.pages.account.classic;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import salesforce.ui.pages.abstracts.account.AccountPageAbstract;
+import salesforce.ui.pages.account.abstracts.AccountPageAbstract;
 import salesforce.utils.DriverMethods;
 
 /**
@@ -26,7 +26,6 @@ import salesforce.utils.DriverMethods;
  * @version 0.0.1
  */
 public class AccountClassicPage extends AccountPageAbstract {
-    //private final String ACCOUNT = "account";
     private String accountList = "//tr[@onmouseout=\"if (window.hiOff){hiOff(this);}\"]//a[contains(text(),\"account\")]";
 
     /**
