+++ src/main/java/salesforce/ui/pages/account/classic/OneAccountClassicPage.java
@@ -11,12 +11,12 @@
  *
  */
 
-package salesforce.ui.pages.classic.account;
+package salesforce.ui.pages.account.classic;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import salesforce.ui.pages.abstracts.account.OneAccountAbstract;
+import salesforce.ui.pages.account.abstracts.OneAccountAbstract;
 
 /**
  * OneAccountClassicPage.
@@ -27,10 +27,13 @@ public class OneAccountClassicPage extends OneAccountAbstract {
 
     @FindBy(xpath = "//td[@class=\" oRight\"]")
     private WebElement accountDetailForm;
+
     @FindBy(xpath = "//div[@class=\"bPageTitle\"]")
     private WebElement accountPanelTitle;
+
     @FindBy(xpath = "//h2[@class=\"topName\"]")
     private WebElement accountTitleLbl;
+
     @FindBy(xpath = "//td[@id=\"topButtonRow\"]//input[@name=\"delete\"]")
     private WebElement deleteBtn;
 
@@ -62,10 +65,9 @@ public class OneAccountClassicPage extends OneAccountAbstract {
 
     /**
      * Delete account.
-     * @param nameAccount string
      */
     @Override
-    public void deleteAccount(final String nameAccount) {
+    public void deleteAccount() {
         deleteBtn.click();
         driver.switchTo().alert().accept();
     }
