+++ src/main/java/salesforce/ui/pages/account/light/AccountLightPage.java
@@ -11,13 +11,13 @@
  *
  */
 
-package salesforce.ui.pages.lightning.account;
+package salesforce.ui.pages.account.light;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import salesforce.ui.pages.abstracts.account.AccountPageAbstract;
+import salesforce.ui.pages.account.abstracts.AccountPageAbstract;
 import salesforce.utils.DriverMethods;
 
 /**
