+++ src/main/java/salesforce/ui/pages/classic/HomeClassicPage.java
@@ -19,7 +19,7 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import salesforce.ui.pages.abstracts.HomePageAbstract;
 import salesforce.ui.pages.abstracts.task.NewTaskAbstract;
 import salesforce.ui.pages.campaign.classic.CampaignClassicPage;
-import salesforce.ui.pages.classic.account.AccountClassicPage;
+import salesforce.ui.pages.account.classic.AccountClassicPage;
 import salesforce.ui.pages.classic.task.NewTaskClassic;
 
 /**
