+++ src/test/java/runner/RunCukesTest.java
@@ -11,7 +11,6 @@ import static runner.GenerateReport.generateMasterReport;
  * RunCukesTest class.
  */
 @CucumberOptions(
-        //plugin = {"pretty", "html:target/cucumber", "json:target/cucumber.json"},
         plugin = {"pretty", "html:target/cucumber-html-report", "json:target/cucumber.json"},
         glue = {"steps", "hooks"},
         features = {"src/test/resources/feature/Account.feature"},
